+++ src/Reporter/DefaultReporter.php
@@ -17,7 +17,6 @@ use Ergebnis\PHPUnit\SlowTestDetector\Comparator;
 use Ergebnis\PHPUnit\SlowTestDetector\Duration;
 use Ergebnis\PHPUnit\SlowTestDetector\Formatter;
 use Ergebnis\PHPUnit\SlowTestDetector\MaximumCount;
-use Ergebnis\PHPUnit\SlowTestDetector\MaximumDuration;
 use Ergebnis\PHPUnit\SlowTestDetector\SlowTest;
 
 /**
@@ -29,7 +28,7 @@ final class DefaultReporter implements Reporter
 
     public function __construct(
         private readonly Formatter\DurationFormatter $durationFormatter,
-        private readonly MaximumDuration $maximumDuration,
+        private readonly Duration $maximumDuration,
         private readonly MaximumCount $maximumCount,
     ) {
         $this->durationComparator = new Comparator\DurationComparator();
@@ -105,7 +104,7 @@ TXT;
 
                 return $maximumDuration;
             },
-            $this->maximumDuration->toDuration(),
+            $this->maximumDuration,
         );
 
         $durationFormatter = $this->durationFormatter;